% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_assembly_tool.R
\name{assemble_output_table}
\alias{assemble_output_table}
\title{Create the new table and populate it. Records are appended
to the new table for every combination of population and
measure.}
\usage{
assemble_output_table(
  population_table,
  measure_table,
  output_database,
  output_schema,
  output_table,
  control_development_mode,
  control_overwrite_output_table,
  control_verbose
)
}
\description{
Create the new table and populate it. Records are appended
to the new table for every combination of population and
measure.
}
