% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_support_helpers.R
\name{handle_summary_case}
\alias{handle_summary_case}
\title{Handle summary function cases}
\usage{
handle_summary_case(
  summary_type,
  proportional,
  m_label,
  m_value,
  m_start_date,
  m_end_date,
  p_start_date,
  p_end_date,
  sqlite = FALSE
)
}
\arguments{
\item{summary_type}{one of the accepted summary keywords}

\item{proportional}{T/F indicator for whether partial overlaps should be
scaled down proportionately.}

\item{m_label}{the (source) label for the measure}

\item{m_value}{the (source) value for the measure}

\item{m_start_date}{the start date for the measure spell}

\item{m_end_date}{the end date for the measure spell}

\item{p_start_date}{the start date for the study period}

\item{p_end_date}{the end date for the study period}

\item{sqlite}{T/F is the connection SQLite (different handling is required
as sqlite does not have CONCAT or DATE-type)}
}
\value{
a list containing three arguments: label, value, and group.
}
\description{
A range of summary functions can be chosen for the output
e.g. MIN, MAX, SUM, COUNT, EXISTS, DURATION, HISTOGRAM, DISTINCT, MEAN.
}
\details{
Each of these needs to be translated into 3 outputs
\enumerate{
\item label - the text label for the result
\item value - the numeric value of the result
\item group - which columns the table needs to be grouped by
These are used to build the SQL code required for assembly.
}
}
