% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_support_helpers.R
\name{has_internal_delimiters}
\alias{has_internal_delimiters}
\title{Checks if internal delimiters exist

Tool inputs are delimited in either [] or "". We require that these inputs
do not contain additional delimiters within the string. For example:
"text\"more text" is not permitted even though the internal delimiter is
escaped.}
\usage{
has_internal_delimiters(string)
}
\arguments{
\item{string}{the character string to check}
}
\value{
T/F whether the input contains internal delimiters
}
\description{
Checks if internal delimiters exist

Tool inputs are delimited in either [] or "". We require that these inputs
do not contain additional delimiters within the string. For example:
"text\"more text" is not permitted even though the internal delimiter is
escaped.
}
