% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assembly.R
\name{needs_work}
\alias{needs_work}
\title{Description: Run the Tool Accelerating Dataset Assembly}
\usage{
needs_work()
}
\description{
Input: User parameters
}
\details{
Output: Dataset suitable for subsequent analysis

Author: Simon Anastasiadis

Dependencies: general_assembly_tool.R, dbplyr_helper_functions.R, utility_functions.R

Notes: runtime 10-30 minutes per cross-section

Issues:

History (reverse order):
2020-11-18 SA v2 for release
2020-05-12 SA v1
2020-01-08 SA v0
#############################################################################
}
