% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_support_helpers.R
\name{prep_for_sql}
\alias{prep_for_sql}
\title{Prepare user input for inclusion in SQL}
\usage{
prep_for_sql(string, alias)
}
\arguments{
\item{string}{character string to prepare}

\item{alias}{table alias to append if needed}
}
\value{
prepared character string
}
\description{
For text input, replaces double " quote delimited strings
with single ' quote delimited strings.
}
\details{
For SQL input, appends alias.
Leaves all other strings unchanged.
}
