% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_assembly_tool.R
\name{prepare_rectangular_table}
\alias{prepare_rectangular_table}
\title{Prepare rectangular table}
\usage{
prepare_rectangular_table(
  population_control_table,
  measures_control_table,
  db_connection,
  output_database,
  output_schema,
  output_table_long,
  output_table_rectangular,
  control_development_mode,
  control_run_checks_only,
  control_silence_progress,
  control_append_long_thin,
  query_path = NA
)
}
\arguments{
\item{population_control_table}{Either the file path to the population and
period control file (in a csv or Excel format) or a data.frame containing
the contents of the control file.}

\item{measures_control_table}{Either the file path to the measures
control file (in a csv or Excel format) or a data.frame containing
the contents of the control file.}

\item{db_connection}{An open connection to the database that contains the
input tables (as specified in the control files) and the output location.}

\item{output_database}{The name of the database that the output should be
saved to. You must have write permissions to this location.
Ideally delimited in square brackets.
Use \code{"[]"} if no database name required.}

\item{output_schema}{The name of the schema that the output should be saved
to. Use \code{"[]"} if the database does not permit schemas. You
must have write permissions to this location.
Ideally delimited in square brackets.}

\item{output_table_long}{The name of the long-thin table that the output
should be saved into. Will be overwritten unless \code{control_append_long_thin}
is set to TRUE. Ideally delimited in square brackets.}

\item{output_table_rectangular}{The name of the rectangular, research-ready
table that the output should be saved into. Will be overwritten if it already
exists. Ideally delimited in square brackets.}

\item{control_development_mode}{T/F should the tool be run in development
mode. In development mode only the first 1000 records for each measure are
assembled. This allows for more rapid testing of correctness.}

\item{control_run_checks_only}{(Optional) T/F should the tool only validate
the control files and not run any database processing. Defaults to FALSE.}

\item{control_silence_progress}{(Optional) T/F should the tool avoid showing
progress information during execution. Defaults to FALSE.}

\item{control_append_long_thin}{(Optional) T/F should the tool append the new
records to the existing long-thin table. Useful if undertaking assembly in
multiple steps. Defaults to FALSE.}

\item{query_path}{If provided will attempt to save a copy of the SQL code
sent to/executed on the database to the provided folder. Save occurs before
execution, hence useful for debugging.}
}
\value{
None, makes permanent changes to the database.
}
\description{
Sub-function of \code{dataset_assembly_tool}. Checks the long-thin table and
pivots to the rectangular, research-ready table.
}
\details{
All inputs as per the master function.
}
