% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_support_helpers.R
\name{prop_warn}
\alias{prop_warn}
\title{Warn if proportional unused}
\usage{
prop_warn(proportional, keyword)
}
\arguments{
\item{proportional}{T/F whether to warn}

\item{keyword}{the matching keyword to warn about}
}
\value{
Nothing, displays warning if proportional is true.
}
\description{
Some summary types can not use the proportional setting.
}
\details{
This function exists to simplify \code{handle_summary_case}
}
