% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_provide_examples.R
\name{provide_assembly_tool_example}
\alias{provide_assembly_tool_example}
\title{Provide an example of the use of the assembly tool}
\usage{
provide_assembly_tool_example(folder = NA)
}
\arguments{
\item{folder}{Folder to copy the example into. Defaults to current working
directory.}
}
\value{
The path of the newly created folder (invisibly)
}
\description{
Provide an example of the use of the assembly tool
}
