% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_support_helpers.R
\name{read_table_file}
\alias{read_table_file}
\title{Auto-detecting which file read command is needed}
\usage{
read_table_file(file_name_and_path)
}
\arguments{
\item{file_name_and_path}{location of the file to read into R}
}
\value{
the file contents as a data.frame
}
\description{
Used to support csv, xls, and xlsx formatted control files.
}
