% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_assembly_tool.R
\docType{data}
\name{required_packages}
\alias{required_packages}
\title{Description: Data assembly tool}
\format{
An object of class \code{character} of length 6.
}
\usage{
required_packages
}
\description{
Input: Two control files (1) population definition, (2) measures for analysis
}
\details{
Output: Summarised measures for each population

Author: Simon Anastasiadis

Dependencies: utility_functions.R, general_assembly_tool_functions.R already sourced

Notes:
\itemize{
\item Uses code folding by headers (Alt + O to collapse all)
}

Issues:

History (reverse order):
2020-11-18 SA v2 for release
2020-01-06 SA v1
2019-11-26 SA v0
#############################################################################
}
\keyword{datasets}
