% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_test_sql_server.R
\name{test_with_sql_server}
\alias{test_with_sql_server}
\title{Run SQL Server-specific tests}
\usage{
test_with_sql_server(
  connection_string,
  table_db = "[IDI_Sandpit]",
  view_db = "[IDI_UserCode]",
  our_schema,
  query_path = NA
)
}
\arguments{
\item{connection_string}{The connection string to connect to the database.
See \code{display_connection_guidance} for guidance on its creation. It is
recommended that the connection be to the database in \code{view_db}.}

\item{table_db}{The name of the database to test writing to.}

\item{view_db}{The name of the database to test creating views in.}

\item{our_schema}{The name of the schema in both databases for testing.}

\item{query_path}{If provided will attempt to save a copy of the SQL code
sent to/executed on the database to the provided folder. Save occurs before
execution, hence useful for debugging.
Due to how testing is implemented, must be an absolute path (not a relative
path) for this function.}
}
\description{
The original version of this package was developed using SQL Server within
the data lab environment. As we do not have an SQL Server available for
this redevelopment we have used SQLite for our testing.
}
\details{
This function provides a way to run many of the original tests within an
SQL Server environment.
}
