% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_validation.R
\name{validate_control_table}
\alias{validate_control_table}
\title{Validate control table - shared validation}
\usage{
validate_control_table(control_table, type, column_reqs)
}
\arguments{
\item{control_table}{the control table to validate}

\item{type}{a string to indicate source control file. Either population or
measure.}

\item{column_reqs}{a data.frame that contains the column requirements for
the table to check.}
}
\value{
T/F whether any check has been failed
}
\description{
Prior to executing the assembly tool, we validate all the inputs to minimize
the chance that the tool fails during data manipulation.
}
\details{
This function does the shared checks for both population and measure control
tables: required columns, no missing values, correct delimiters, no internal
delimiters.
}
