% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_validation.R
\name{validate_database_tables}
\alias{validate_database_tables}
\title{Validate database tables}
\usage{
validate_database_tables(control_table, db_connection)
}
\arguments{
\item{control_table}{a control table to validate}

\item{db_connection}{a connection to the database to validate against}
}
\value{
T/F whether any check has been failed
}
\description{
Check that all tables requested in control tables exist
and that their required columns exist
}
