% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_validation.R
\name{validate_measure_control_table_2}
\alias{validate_measure_control_table_2}
\title{Validate measure control table 2}
\usage{
validate_measure_control_table_2(measure_table)
}
\arguments{
\item{measure_table}{the measure control table to validate}
}
\value{
T/F whether any check has been failed
}
\description{
Prior to executing the assembly tool, we validate all the inputs to minimize
the chance that the tool fails during data manipulation.
}
\details{
Check the measure control table, with the checks that are specific to this
control table:
}
