% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_validation.R
\name{validate_populaiton_control_table}
\alias{validate_populaiton_control_table}
\title{Validate population and period control table}
\usage{
validate_populaiton_control_table(population_table)
}
\arguments{
\item{population_table}{the population control table to validate}
}
\value{
T/F whether any check has been failed
}
\description{
Prior to executing the assembly tool, we validate all the inputs to minimize
the chance that the tool fails during data manipulation.
}
\details{
Check the population control table, with the checks that are shared with the
measure control table.
}
